update Classes Entity
Use Hibernate Tools to Create Entity from Tables in database

+++ src/main/java/com/sonnguyen/webdochoioto/controller/RestUserController.java
@@ -29,7 +29,7 @@ public class RestUserController {
 	}
 	
 	@RequestMapping(value="/user/{userId}",method=RequestMethod.GET)
-	public  UserDTO viewOneUser(HttpServletRequest request,@PathVariable ("userId") Long id) {
+	public  UserDTO viewOneUser(HttpServletRequest request,@PathVariable ("userId") Integer id) {
 		return userService.getUserById(id);
 	}
