+++ src/main/java/com/sonnguyen/webdochoioto/service/UserService.java
@@ -7,7 +7,7 @@ import com.sonnguyen.webdochoioto.model.UserDTO;
 public interface UserService {
 	public void insert(UserDTO user);
 	public void update(UserDTO user);
-	public void delete(Long id);
-	public UserDTO getUserById(Long id);
+	public void delete(Integer id);
+	public UserDTO getUserById(Integer id);
 	public List<UserDTO> getAllUser();
 }
